Restructuring class model, for reusability purpose, BugFixing
