+++ src/Lithnet.ResourceManagement.Automation/AttributeValueExtension.cs
@@ -12,6 +12,8 @@ namespace Lithnet.ResourceManagement.Automation
         private const string FILTERHEADER = @"Dialect=""http://schemas.microsoft.com/2006/11/XPathFilterDialect""";
         public static bool IsFilter(this AttributeValue AttributeValue)
         {
+            if (AttributeValue.Attribute.IsMultivalued)
+                return false;
             return AttributeValue.StringValue.Contains(FILTERHEADER);
         }        
     }
