--- src/Lithnet.ResourceManagement.Automation/ConfigSyncController/ConfigSyncController.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Lithnet.ResourceManagement.Automation
-{
-    public class ConfigSyncController
-    {
-        public List<ConfigSyncConfiguration> Configurations { get; set; }
-    }
-}
