--- src/Lithnet.ResourceManagement.Automation/ConfigSyncController/ObjectExclusion.cs
@@ -1,15 +1,0 @@
-﻿using System;
-using System.Collections;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Lithnet.ResourceManagement.Automation
-{
-    public class ObjectExclusion
-    {
-        public Dictionary<string,string> AnchorKeyValueList { get; set; }
-        public List<string> AttributExclusions { get; set; }
-    }
-}
