+++ src/Lithnet.ResourceManagement.Automation/Lithnet.ResourceManagement.Automation.csproj
@@ -35,6 +35,9 @@
     <Prefer32Bit>false</Prefer32Bit>
   </PropertyGroup>
   <ItemGroup>
+    <Reference Include="JetBrains.Annotations, Version=11.1.0.0, Culture=neutral, PublicKeyToken=1010a0d8d6380325, processorArchitecture=MSIL">
+      <HintPath>..\packages\JetBrains.Annotations.11.1.0\lib\net20\JetBrains.Annotations.dll</HintPath>
+    </Reference>
     <Reference Include="Lithnet.ResourceManagement.Client, Version=1.0.6623.18312, Culture=neutral, PublicKeyToken=bd0636700c1e2538, processorArchitecture=MSIL">
       <HintPath>..\packages\Lithnet.ResourceManagement.Client.1.0.6623.18312\lib\net40\Lithnet.ResourceManagement.Client.dll</HintPath>
     </Reference>
@@ -62,26 +65,26 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="AsyncCmdlet.cs" />
     <Compile Include="AttributeValueArrayList.cs" />
     <Compile Include="AttributeValueExtension.cs" />
-    <Compile Include="ConfigSyncController\ConfigSyncController.cs" />
+    <Compile Include="ChangeObserver\ChangeModeType.cs" />
+    <Compile Include="ChangeObserver\ObserverObjectSetting.cs" />
+    <Compile Include="ChangeObserver\ObserverSetting.cs" />
+    <Compile Include="ChangeObserver\StartChangeObserver.cs" />
+    <Compile Include="RMConfigConverter\ConverterSetting.cs" />
     <Compile Include="ConfigSyncControl.cs" />
-    <Compile Include="ConfigSyncController\ConfigSyncConfiguration.cs" />
-    <Compile Include="ConfigSyncController\ImportConfigSyncController.cs" />
-    <Compile Include="ConfigSyncController\NewConfigSyncConfiguration.cs" />
-    <Compile Include="ConfigSyncController\NewObjectExclusion.cs" />
-    <Compile Include="ConfigSyncController\SaveConfigSyncController.cs" />
-    <Compile Include="ConvertToRMConfig.cs" />
+    <Compile Include="RMConfigConverter\ObjectSetting.cs" />
+    <Compile Include="RMConfigConverter\ConvertToRMConfig.cs" />
     <Compile Include="Enums\AttributeOperationType.cs" />
     <Compile Include="Enums\ApprovalDecision.cs" />
     <Compile Include="Enums\AttributeValueType.cs" />
     <Compile Include="Enums\SchemaRefreshEvent.cs" />
     <Compile Include="Enums\ResourceOperationType.cs" />
     <Compile Include="Enums\WellKnownGuids.cs" />
-    <Compile Include="ConfigSyncController\ObjectExclusion.cs" />
-    <Compile Include="ConfigSyncController\NewConfigSyncController.cs" />
-    <Compile Include="OutRMConfig.cs" />
-    <Compile Include="ResourceOperationExtension.cs" />
+    <Compile Include="RMConfigConverter\ObjectExclusion.cs" />
+    <Compile Include="RMConfigConverter\OutRMConfig.cs" />
+    <Compile Include="RMConfigConverter\Converter.cs" />
     <Compile Include="SetPendingApprovalRequest.cs" />
     <Compile Include="GetApprovalRequest.cs" />
     <Compile Include="ImportRMConfig.cs" />
