+++ src/Lithnet.ResourceManagement.Automation/RMConfigConverter/ObjectSetting.cs
@@ -5,11 +5,29 @@ using System.Management.Automation;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Lithnet.ResourceManagement.Automation
+namespace Lithnet.ResourceManagement.Automation.RMConfigConverter
 {
+    public class ObjectSetting
+    {
+        public string ObjectType { get; set; }
+        public List<string> AnchorAttributes { get; set; }
 
-    [Cmdlet(VerbsCommon.New, "ConfigSyncConfiguration")]    
-    public class NewConfigSyncConfiguration : PSCmdlet
+        public string IDPrefix { get; set; }
+        
+        public bool ReferenceResolution { get; set; }
+        public List<string> ReferenceResolutionAttributExclusions { get; set; }
+
+        public bool IncludeDefaultAttributes { get; set; }
+
+        public bool IncludeEmptyAttributeValues { get; set; }
+
+        public List<string> AttributExclusions { get; set; }
+
+        public List<ObjectExclusion> ObjectSpecificExlusions { get; set; }        
+    }
+
+    [Cmdlet(VerbsCommon.New, "RMConverterObjectSetting")]
+    public class NewObjectSetting : PSCmdlet
     {
         [Parameter(Mandatory = true, Position = 1)]
         public string ObjectType { get; set; }
@@ -19,36 +37,40 @@ namespace Lithnet.ResourceManagement.Automation
 
         [Parameter(Mandatory = false, Position = 3)]
         public string IDPrefix { get; set; }
-        
+
         [Parameter(Mandatory = false, Position = 4)]
         public SwitchParameter ReferenceResolution { get; set; }
 
         [Parameter(Mandatory = false, Position = 5)]
-        public SwitchParameter IncludeDefaultAttributes { get; set; }
+        public string[] ReferenceResolutionAttributExclusions { get; set; }
 
         [Parameter(Mandatory = false, Position = 6)]
-        public SwitchParameter IncludeEmptyAttributeValues { get; set; }
+        public SwitchParameter IncludeDefaultAttributes { get; set; }
 
         [Parameter(Mandatory = false, Position = 7)]
-        public List<string> AttributExclusions { get; set; }
+        public SwitchParameter IncludeEmptyAttributeValues { get; set; }
 
         [Parameter(Mandatory = false, Position = 8)]
+        public List<string> AttributExclusions { get; set; }
+
+        [Parameter(Mandatory = false, Position = 9)]
         public ObjectExclusion[] ObjectSpecificExlusions { get; set; }
 
 
         protected override void ProcessRecord()
         {
             this.WriteObject(
-                new ConfigSyncConfiguration()
+                new ObjectSetting()
                 {
                     AnchorAttributes = AnchorAttributes?.ToList(),
                     AttributExclusions = AttributExclusions?.ToList(),
-                    IDPrefix = IDPrefix,                    
+                    IDPrefix = IDPrefix,
                     IncludeDefaultAttributes = IncludeDefaultAttributes.IsPresent,
                     IncludeEmptyAttributeValues = IncludeEmptyAttributeValues.IsPresent,
                     ObjectSpecificExlusions = ObjectSpecificExlusions?.ToList(),
                     ObjectType = ObjectType,
-                    ReferenceResolution = ReferenceResolution.IsPresent
+                    ReferenceResolution = ReferenceResolution.IsPresent,
+                    ReferenceResolutionAttributExclusions = ReferenceResolutionAttributExclusions?.ToList()
                 });
         }
     }
